+++ sd_scripts/PaddleSeg/paddleseg/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import models, datasets, transforms, optimizers
+# from . import models, datasets, transforms, optimizers
 
 __version__ = '2.8.0'
