--- sd_scripts/library/face_tool/.gitignore
@@ -1,44 +1,0 @@
-__pycache__
-*.ckpt
-*.safetensors
-*.pth
-/ESRGAN/*
-/SwinIR/*
-/repositories
-/venv
-/tmp
-/model.ckpt
-/models/**/*
-/GFPGANv1.3.pth
-/gfpgan/weights/*.pth
-/ui-config.json
-/outputs
-/config.json
-/log
-/webui.settings.bat
-/embeddings
-/styles.csv
-/params.txt
-/styles.csv.bak
-/webui-user.bat
-/webui-user.sh
-/interrogate
-/user.css
-/.idea
-notification.mp3
-/SwinIR
-/textual_inversion
-.vscode
-/extensions
-/test/stdout.txt
-/test/stderr.txt
-/cache.json*
-/config_states/
-/cache.json
-cache.json.lock
-gradio/launches.json
-models/
-stable-diffusion-webui/
-/node_modules
-/package-lock.json
-/.coverage*
