--- sd_scripts/library/face_tool/app/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Time    : 2023/9/8 11:03 AM
-# @Author  : wangdongming
-# @Site    : 
-# @File    : __init__.py
-# @Software: Hifive
