--- sd_scripts/library/face_tool/app/routers/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Time    : 2023/9/8 10:54 AM
-# @Author  : wangdongming
-# @Site    : 
-# @File    : __init__.py.py
-# @Software: Hifive
