--- sd_scripts/library/face_tool/data_models/types.py
@@ -1,13 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Time    : 2023/9/7 8:37 PM
-# @Author  : wangdongming
-# @Site    : 
-# @File    : types.py
-# @Software: Hifive
-
-BaseRoutePath = '/tools'
-V1 = '/v1'
-V2 = '/v2'
-Face = '/face'
-SDModel = '/sd-model'
