--- sd_scripts/library/face_tool/requirements.txt
@@ -1,72 +1,0 @@
-albumentations==1.3.1
-aliyun-python-sdk-core==2.13.36
-aliyun-python-sdk-kms==2.16.1
-annotated-types==0.5.0
-anyio==3.7.1
-APScheduler==3.10.4
-#backports.zoneinfo==0.2.1
-certifi==2023.7.22
-cffi==1.15.1
-charset-normalizer==3.2.0
-click==8.1.7
-coloredlogs==15.0.1
-crcmod==1.7
-cryptography==41.0.3
-cycler==0.11.0
-Cython==3.0.2
-easydict==1.10
-esdk-obs-python==3.22.2
-exceptiongroup==1.1.3
-fastapi==0.103.1
-flatbuffers==23.5.26
-fonttools==4.38.0
-h11==0.14.0
-humanfriendly==10.0
-idna==3.4
-imageio==2.31.2
-importlib-metadata==6.7.0
-insightface==0.7.3
-jmespath==0.10.0
-joblib==1.3.2
-kiwisolver==1.4.5
-loguru==0.7.1
-matplotlib==3.5.3
-mpmath==1.3.0
-networkx==2.6.3
-numpy
-onnx==1.14.1
-onnxruntime==1.14.1
-opencv-python==4.8.0.76
-opencv-python-headless==4.8.0.76
-oss2==2.18.1
-packaging==23.1
-Pillow==9.5.0
-prettytable==3.7.0
-protobuf==4.24.2
-pycparser==2.21
-pycryptodome==3.10.1
-pydantic==2.3.0
-pydantic_core==2.6.3
-pyparsing==3.1.1
-python-dateutil==2.8.2
-pytz==2023.3.post1
-PyWavelets==1.3.0
-PyYAML==6.0.1
-qudida==0.0.4
-requests==2.31.0
-scikit-image==0.19.3
-scikit-learn==1.0.2
-scipy==1.7.3
-six==1.16.0
-sniffio==1.3.0
-starlette==0.27.0
-sympy==1.10.1
-threadpoolctl==3.1.0
-tifffile==2021.11.2
-tqdm==4.66.1
-typing_extensions==4.7.1
-tzlocal==5.0.1
-urllib3==2.0.4
-uvicorn==0.22.0
-wcwidth==0.2.6
-zipp==3.15.0
