--- sd_scripts/library/face_tool/sources.list
@@ -1,8 +1,0 @@
-deb https://mirrors.aliyun.com/debian/ bookworm main non-free non-free-firmware contrib
-deb-src https://mirrors.aliyun.com/debian/ bookworm main non-free non-free-firmware contrib
-deb https://mirrors.aliyun.com/debian-security/ bookworm-security main
-deb-src https://mirrors.aliyun.com/debian-security/ bookworm-security main
-deb https://mirrors.aliyun.com/debian/ bookworm-updates main non-free non-free-firmware contrib
-deb-src https://mirrors.aliyun.com/debian/ bookworm-updates main non-free non-free-firmware contrib
-deb https://mirrors.aliyun.com/debian/ bookworm-backports main non-free non-free-firmware contrib
-deb-src https://mirrors.aliyun.com/debian/ bookworm-backports main non-free non-free-firmware contrib
