--- sd_scripts/library/face_tool/super/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Time    : 2023/9/7 5:28 PM
-# @Author  : wangdongming
-# @Site    : 
-# @File    : __init__.py.py
-# @Software: Hifive
