+++ sd_scripts/train_auto_xz.py
@@ -23,7 +23,7 @@ import argparse
 
 import sys
 
-sys.path.append("PaddleSeg/contrib/PP-HumanSeg")
+sys.path.append("sd_scripts/PaddleSeg/contrib/PP-HumanSeg")
 from src.seg_demo import seg_image
 
 import sd_scripts.library.config_util as config_util
@@ -38,8 +38,8 @@ from modelscope.utils.constant import Tasks
 from modelscope import snapshot_download
 from transformers import PreTrainedTokenizerBase,PreTrainedModel
 
-from library.transformers_pretrained import ori_tokenizer_from_pretrained,ori_model_from_pretrained
-from library.face_tool.super.face import insightface_main_face
+from sd_scripts.library.transformers_pretrained import ori_tokenizer_from_pretrained,ori_model_from_pretrained
+from sd_scripts.library.face_tool import insightface_main_face
 
 
 def patch_tokenizer_base():
